Move functions back from utils to module.
